% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AnnualizedReturn.R
\name{AnnualizedReturn}
\alias{AnnualizedReturn}
\title{This function is used in the computation of the  annulized return of a bond
once the return cashflow matrix is computed}
\usage{
AnnualizedReturn(BeginValue = numeric(), EndValue = numeric(),
  TimeHorizon = numeric(), Frequency = numeric())
}
\arguments{
\item{BeginValue}{the beginning value of the portfolio position}

\item{EndValue}{the ending value of the portfolio position including coupon income, reinvestment, and price change}

\item{TimeHorizon}{A numeric value the time horizon over which to calculate the return}

\item{Frequency}{the payment frequency of the bond}
}
\description{
This function is used in the computation of the  annulized return of a bond
once the return cashflow matrix is computed
}

