% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Bondprice.R
\name{bond.irr}
\alias{bond.irr}
\title{bond.irr is an intermediate function to compute bond yield to maturity}
\usage{
bond.irr(yield.to.maturity, coupon, coupon.frequency, years.mat,
  price = price, face.value)
}
\arguments{
\item{yield.to.maturity}{a numeric value the yield to maturity}

\item{coupon}{a numeric value the bond coupon}

\item{coupon.frequency}{a numeric value the frequency of coupon payments}

\item{years.mat}{a numeric value the years to maturity}

\item{price}{a numeric value the price of the bond}

\item{face.value}{a numeric value the face value of the bond}
}
\description{
bond.irr is an intermediate function to compute bond yield to maturity
}

