% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Bondprice.R
\name{bond.price}
\alias{bond.price}
\title{bond.price illustrates the basic compution of the price of a bond}
\usage{
bond.price(yield.to.maturity, coupon, coupon.frequency, years.mat, face.value)
}
\arguments{
\item{yield.to.maturity}{numeric value the yield to maturity of the bond for example 4.0\% entered as 0.04}

\item{coupon}{numeric value the coupon of the bond for example 4.0\% entered as  0.04}

\item{coupon.frequency}{a numeric value the number of coupon payments for example semi-annual entered as 2}

\item{years.mat}{a numeric value remaining years to maturity.  This is a simple function, remaining years must equal frequency}

\item{face.value}{a numeric value the notional or par amount of the bond}
}
\description{
Computation of price as the present value of the coupon annuity payment and zero coupon price of the principal payment at maturity
This function is taken directly from BondLab and is copyright BondLab Technologies, Inc.
}
\examples{
bond.price(yield.to.maturity = .04, coupon = .04, coupon.frequency = 2,
years.mat = 10, face.value = 100)
}

