% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MortgageCredit.R
\name{CreditEnhancement}
\alias{CreditEnhancement}
\title{Function for sizing mortgage credit enhancement}
\usage{
CreditEnhancement(bond.id = character, trade.date = character,
  settlement.date = character, original.balance = numeric(),
  sigma = numeric(), HPA = numeric(), HPV = numeric(),
  paths = numeric())
}
\arguments{
\item{bond.id}{a character string the bond id}

\item{trade.date}{a character string the trade date}

\item{settlement.date}{a character string the settlement date}

\item{original.balance}{a numeric value the original balance}

\item{sigma}{a numeric value the **annualized** volatility assumption used for
the CIR interest rate model}

\item{HPA}{a numeric value the long run expected home price appreciation for example
2.5\% is entered as 0.025}

\item{HPV}{the annualized volatility of home prices for example 1.5\% is entered as 0.015}

\item{paths}{a numeric value the number of paths}
}
\description{
This function sizes mortgage credit enhancement levels.  The funtion illustrates how
one may create a new analytic function by calling individial BondLab functions.  In this case
we create a new function CreditEnhancement
}
\examples{
CreditEnhancement(bond.id = bondlabMBS4, original.balance = 100000,
trade.date = "01-10-2013", settlement.date = "01-13-2013", sigma = .2, HPA = .025,
HPV = .015, paths = 100)
}

