% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ExtractScenario.R
\name{ExtractScenario}
\alias{ExtractScenario}
\title{Extract Scenario}
\usage{
ExtractScenario(bond.id = "character")
}
\arguments{
\item{bond.id}{A character string the name or id assigned to pass-through analytics object}
}
\description{
Fuction to extract scenario results from Bond Lab analytics object.  This is taken directory
from BondLab and is copyright Bond Lab Technologies, Inc.
}

