% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LotteryPresentValue.R
\name{LotteryDiscountRate}
\alias{LotteryDiscountRate}
\title{This function solves for the discount rate of the lottery problem in Chapter 1}
\usage{
LotteryDiscountRate(LumpSumPayment = numeric(), payment = numeric(),
  number.period = numeric(), frequency = numeric())
}
\arguments{
\item{LumpSumPayment}{a numeric value the lump sum paid}

\item{payment}{a numeric value the periodic annuity payment}

\item{number.period}{a numeric value the number of payments}

\item{frequency}{a numeric value the frequency of payments}
}
\description{
This function solves for the discount rate of the lottery problem in Chapter 1
}
\examples{
LotteryDiscountRate(LumpSumPayment = 500000, payment = 2777.78, number.period = 360, frequency = 12)
}

