% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LotteryPresentValue.R
\name{LotteryPresentValue}
\alias{LotteryPresentValue}
\title{This function solves for the present value of a lottery payment in Chapter 1}
\usage{
LotteryPresentValue(payment = numeric(), interest.rate = numeric(),
  number.period = numeric(), frequency = numeric())
}
\arguments{
\item{payment}{The annuity payment received by the winner}

\item{interest.rate}{The discount rate used to determine the present value for example 4.0\% entered as 0.04}

\item{number.period}{The number of periods example the lottery PV problem is 360 periods}

\item{frequency}{The frequency payments the lottert PV problem is 12 frequency (monthly)}
}
\description{
This function solves for the present value of a lottery payment in Chapter 1
}
\examples{
LotteryPresentValue(payment = 2777.78, interest.rate = .0530, number.period = 360, frequency = 12)
}

