% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MortgageAmoritization.R
\name{MortgageAmortization}
\alias{MortgageAmortization}
\title{Fixed rate mortgage amortization schedule}
\usage{
MortgageAmortization(OriginalBal = numeric(), NoteRate = numeric(),
  Term = numeric())
}
\arguments{
\item{OriginalBal}{A numeric value the original loan amount}

\item{NoteRate}{A numeric value the borrower note rate}

\item{Term}{A numeric value the term of the loan in months}
}
\description{
Fixed rate mortgage amortization schedule
}
\examples{
MortgageAmortization(OriginalBal = 200000, NoteRate = 4.5, Term = 360)
}

