% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PassThroughOAS.R
\name{PassThrough.OAS}
\alias{PassThrough.OAS}
\title{OAS Analaysis of Pass Though MBS}
\usage{
PassThrough.OAS(bond.id = character, trade.date = character,
  settlement.date = character, original.balance = numeric(),
  price = numeric(), sigma = numeric(), paths = numeric())
}
\arguments{
\item{bond.id}{a character string the bond id}

\item{trade.date}{a character string the trade date}

\item{settlement.date}{a character string the settlement date}

\item{original.balance}{a numeric value the original balance}

\item{price}{a numeric value the price}

\item{sigma}{a numeric value the **annualized** volatility.  The volatilty assumption assumes}

\item{paths}{a numeric value the number of interest rate paths
a trading year of 252 days}
}
\description{
Function calls BondLab PassThroughOAS and plots the results
copyright Bond Lab Technologies, Inc 2015
}
\examples{
PassThrough.OAS(bond.id = "bondlabMBS4", trade.date = "01-10-2013",
settlement.date = "01-13-2013", original.balance = 100000, price = 105.75, sigma = .20, paths = 200)
}

