% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PassThroughCashFlow.R
\name{PassThroughCashFlow}
\alias{PassThroughCashFlow}
\title{Function plots a mortgage pass-thorugh security cash flow}
\usage{
PassThroughCashFlow(bond.id = character, trade.date = character,
  settlement.date = character, begin.cpr = numeric(), end.cpr = numeric(),
  seasoning.period = numeric())
}
\arguments{
\item{bond.id}{a character string the bond id}

\item{trade.date}{a character string the trade date}

\item{settlement.date}{a character string the settlement date}

\item{begin.cpr}{a numeric value the first month CPR assumption}

\item{end.cpr}{a numeric value end CPR of the seasoning ramp}

\item{seasoning.period}{a numeric value the length of the seasoning ramp}
}
\description{
Function plots a mortgage pass-thorugh security cash flow
}
\examples{
PassThroughCashFlow(bond.id = "bondlabMBS4", trade.date = "01-10-2013",
settlement.date = "01-13-2013", begin.cpr = .002, end.cpr = .06, seasoning.period = 30)
}

