% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MortgageKeyRate.R
\name{PlotMtgKeyRates}
\alias{PlotMtgKeyRates}
\title{Plot the mortgage key rate duration}
\usage{
PlotMtgKeyRates(bond.id = character, original.bal = numeric(),
  price = numeric(), trade.date = character, settlement.date = character)
}
\arguments{
\item{bond.id}{a character vector either bondlabMBS4 or bondlabMBS55}

\item{original.bal}{a numeric value the pool original balance}

\item{price}{a numeric value the price of the bond}

\item{trade.date}{a character value the trade date}

\item{settlement.date}{a character value the settlement date}
}
\description{
The mortgage key rate plot function defaults to PrepaymentAssumption = "MODEL"
and scenario.set = "NC"
}
\examples{
PlotMtgKeyRates(bond.id = "bondlabMBS4", original.bal = 1000,
price = 105.75, trade.date = "01-10-2013", settlement.date = "01-13-2013")
}

