% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_termstructure.R
\name{PlotTermStructure}
\alias{PlotTermStructure}
\title{Plot Term Structure}
\usage{
PlotTermStructure(trade.date = character, method = character)
}
\arguments{
\item{trade.date}{A character value the trade.date}

\item{method}{A character value the method passed to termstrc to fit the term structure suggest "ns"}
}
\description{
A function to plot the term structure of interest rates
}
\examples{
PlotTermStructure(trade.date = "01-10-2013", method = "ns")
}

