% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/timevalue.R
\name{PresentValue}
\alias{PresentValue}
\title{Present Value}
\usage{
PresentValue(interest.rate = numeric(), number.period = numeric(),
  frequency = numeric())
}
\arguments{
\item{interest.rate}{a numeric value the discount rate used entered for example 4.0\% is entered as .04}

\item{number.period}{an numeric value the number of periods over which to discount}

\item{frequency}{a numeric value the frequency of payments made annually}
}
\description{
The function below is a wrapper around the BondLab function TimeValue returning present value
}
\examples{
PresentValue(interest.rate = .05, number.period = 3, frequency = 1)
}

