% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/REMICOAS.R
\name{REMICOAS}
\alias{REMICOAS}
\title{OAS Analaysis of REMIC}
\usage{
REMICOAS(bond.id = character, trade.date = character,
  settlement.date = character, tranche.price = numeric(),
  collateral.price = numeric(), sigma = numeric(), paths = numeric())
}
\arguments{
\item{bond.id}{a character string the bond id}

\item{trade.date}{a character string the trade date}

\item{settlement.date}{a character string the settlement date}

\item{tranche.price}{a numeric value the price}

\item{collateral.price}{a numeric value the price of the underlying collateral}

\item{sigma}{a numeric value the **annualized** volatility.  The volatilty assumption assumes}

\item{paths}{a numeric value the number of interest rate paths
a trading year of 252 days}
}
\description{
Function calls BondLab REMIC.OAS and plots the results
copyright Bond Lab Technologies, Inc 2015
}
\examples{
REMICOAS(bond.id = "BondLabSEQ1", trade.date = "01-10-2013", settlement.date = "01-13-2013",
tranche.price = 100, collateral.price = 100, sigma = 0.20, paths = 200)
}

