% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ReturnAnalysis.R
\name{ReturnAnalysis}
\alias{ReturnAnalysis}
\title{Function illustrates basic return analyis}
\usage{
ReturnAnalysis()
}
\description{
cash flow analysis of fixed income return
 the function replicates Table 5.2
}

