% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MtgTotalReturn.R
\name{TotalReturn}
\alias{TotalReturn}
\title{Mortgage Total Return Graph}
\usage{
TotalReturn()
}
\description{
Calculates and plots mortgage pass-through total return analysis
This function illustrates the call to BondLab PassThroughAnalytics
for the bondlabMBS4 and bondlabMBS55.
}

