% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TwistCaseStudy.R
\name{TwistScenario}
\alias{TwistScenario}
\title{This function illustrates twist scenario analysis}
\usage{
TwistScenario(bond.id = "character", original.bal = numeric(),
  price = numeric(), trade.date = "character",
  settlement.date = "character", method = "ns")
}
\arguments{
\item{bond.id}{a character string the bond id}

\item{original.bal}{a numeric value the original balance}

\item{price}{a numeric value the price}

\item{trade.date}{a character string the trade date}

\item{settlement.date}{a character string the settlement date}

\item{method}{a character the method used to fit the term strucure use method "ns"}
}
\description{
TwistScenario plots the expected horizon return under each scenario.
TwistScenario is wrapper around PassThroughAnalytics see
the BondLab help file for more detail.
}
\examples{
TwistScenario(bond.id = "bondlabMBS4", original.bal = 100000,
price = 105.75, trade.date = "01-10-2013", settlement.date = "01-13-2013", method = "ns")
}

