% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FrameworkPlot.R
\name{ValuationFramework}
\alias{ValuationFramework}
\title{Structured Securities Valuation Framework Graphic}
\usage{
ValuationFramework()
}
\description{
The graphic for the valuation of structured securities
}

