% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Bondprice.R
\name{YieldToMaturity}
\alias{YieldToMaturity}
\title{YTM calculates the bond yield to maturity}
\usage{
YieldToMaturity(coupon = numeric(), coupon.frequency = numeric(),
  years.mat = numeric(), price = numeric())
}
\arguments{
\item{coupon}{a numeric value the coupon rate example 5.0\% is entered as 0.05}

\item{coupon.frequency}{a numeric value the annual payment frequency}

\item{years.mat}{a numeric value the years to maturity}

\item{price}{a numeric value the price}
}
\description{
The function calls uniroot and bond.irr
}

